shift toward standard layout for package oriented design
https://www.goinggo.net/2017/02/package-oriented-design.html
