+++ cmd/arachneserv/arachne2pb.go
@@ -3,7 +3,7 @@ package main
 import (
 	//	"context"
 
-	"github.com/dougfort/arachne/game"
+	"github.com/dougfort/arachne/internal/game"
 
 	pb "github.com/dougfort/arachne/arachne"
 )
