+++ cmd/arachneserv/start_game.go
@@ -6,35 +6,37 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/dougfort/arachne/game"
+	"github.com/dougfort/arachne/internal/game"
 
 	pb "github.com/dougfort/arachne/arachne"
 )
 
-// RequestDeal requests a deal
-func (s *arachneServer) RequestDeal(
+// StartGame starts a new game
+func (s *arachneServer) StartGame(
 	ctx context.Context,
-	request *pb.DealRequest,
+	request *pb.GameRequest,
 ) (*pb.Game, error) {
 	var localGame *game.Game
 	var pbGame pb.Game
-	var ok bool
 
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	if localGame, ok = s.active[request.GetId()]; !ok {
-		return nil, errors.Errorf("unknown game id %d", request.GetId())
-	}
-
-	if err := localGame.Deal(); err != nil {
-		return nil, errors.Wrap(err, "Deal failed")
+	switch request.Gametype {
+	case pb.GameRequest_RANDOM:
+		localGame = game.New()
+	case pb.GameRequest_REPLAY:
+		localGame = game.Replay(request.Seed)
+	default:
+		return nil, errors.Errorf("invalid GameType %d", request.Gametype)
 	}
 
 	pbGame.Seed = localGame.Deck.Seed()
 	pbGame.Stack = arachne2pb(localGame.Tableau)
 	pbGame.CardsRemaining = int32(localGame.Deck.RemainingCards())
 
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+
+	pbGame.Id = s.nextID
+	s.nextID++
 	s.active[pbGame.Id] = localGame
 	return &pbGame, nil
 }
