+++ cmd/botclient/doc.go
