+++ cmd/textclient/doc.go
