+++ install_textclient.sh
@@ -4,6 +4,6 @@
 set -e
 set -x
 
-pushd $GOPATH/src/github.com/dougfort/arachne/textclient
+pushd $GOPATH/src/github.com/dougfort/arachne/cmd/textclient
 go install -race
 popd
