+++ internal/game/types.go
