+++ internal/kit/doc.go
